Update Skills to Heroes 1.5.3 and Spigot

+++ src/com/herocraftonline/heroes/characters/skill/skills/SkillAbsolution.java
@@ -46,8 +46,8 @@ public class SkillAbsolution extends TargettedSkill {
         }
 
         Hero targetHero = plugin.getCharacterManager().getHero((Player) target);
-        int hpPlus = SkillConfigManager.getUseSetting(hero, this, SkillSetting.HEALTH, 10, false);
-        int targetHealth = target.getHealth();
+        double hpPlus = SkillConfigManager.getUseSetting(hero, this, SkillSetting.HEALTH, 10, false);
+        double targetHealth = target.getHealth();
 
         if (targetHealth >= target.getMaxHealth()) {
             if (player.equals(targetHero.getPlayer())) {
