+++ src/com/herocraftonline/heroes/characters/skill/skills/SkillChaosOrb.java
@@ -113,7 +113,7 @@ public class SkillChaosOrb extends ActiveSkill {
 
                 // Damage the player
                 addSpellTarget(entity, hero);
-                int damage = SkillConfigManager.getUseSetting(hero, skill, SkillSetting.DAMAGE, 4, false);
+                double damage = SkillConfigManager.getUseSetting(hero, skill, SkillSetting.DAMAGE, 4, false);
                 damage += (int) (SkillConfigManager.getUseSetting(hero, skill, SkillSetting.DAMAGE_INCREASE, 0.0, false) * hero.getSkillLevel(skill));
                 damageEntity(entity, hero.getPlayer(), damage, EntityDamageEvent.DamageCause.MAGIC);
                 event.setCancelled(true);
