+++ src/com/herocraftonline/heroes/characters/skill/skills/SkillConviction.java
@@ -147,7 +147,7 @@ public class SkillConviction extends ActiveSkill {
             CharacterTemplate character = SkillConviction.this.plugin.getCharacterManager().getCharacter((LivingEntity) event.getEntity());
             if (character.hasEffect("Conviction")) {
                 double damageModifier = ((ConvictionEffect) character.getEffect("Conviction")).damageModifier;
-                event.setDamage((int) (event.getDamage() * damageModifier));
+                event.setDamage((event.getDamage() * damageModifier));
             }           
         }
         
@@ -157,7 +157,7 @@ public class SkillConviction extends ActiveSkill {
         	CharacterTemplate character = SkillConviction.this.plugin.getCharacterManager().getCharacter((LivingEntity) event.getEntity());
         	if(character.hasEffect("Conviction")) {
         		double damageModifier = ((ConvictionEffect) character.getEffect("Conviction")).damageModifier;
-        		event.setDamage((int)(event.getDamage() * damageModifier));
+        		event.setDamage((event.getDamage() * damageModifier));
         	}
         }
     }
