+++ src/com/herocraftonline/heroes/characters/skill/skills/SkillDivineWord.java
@@ -45,8 +45,8 @@ public class SkillDivineWord extends TargettedSkill {
         }
 
         Hero targetHero = plugin.getCharacterManager().getHero((Player) target);
-        int hpPlus = SkillConfigManager.getUseSetting(hero, this, SkillSetting.HEALTH, 5, false);
-        int targetHealth = target.getHealth();
+        double hpPlus = SkillConfigManager.getUseSetting(hero, this, SkillSetting.HEALTH, 5, false);
+        double targetHealth = target.getHealth();
 
         if (targetHealth >= target.getMaxHealth()) {
             if (player.equals(targetHero.getPlayer())) {
