+++ src/com/herocraftonline/heroes/characters/skill/skills/SkillEntangle.java
@@ -74,7 +74,7 @@ public class SkillEntangle extends TargettedSkill {
     public SkillResult use(Hero hero, LivingEntity target, String[] args) {
 
         //deal  damage
-        int damage = SkillConfigManager.getUseSetting(hero, this, SkillSetting.DAMAGE, 1, false);
+        double damage = SkillConfigManager.getUseSetting(hero, this, SkillSetting.DAMAGE, 1, false);
         damageEntity(target, hero.getPlayer(), damage, EntityDamageEvent.DamageCause.MAGIC);
 
         int duration = SkillConfigManager.getUseSetting(hero, this, SkillSetting.DURATION, 4000, false);
