+++ src/com/herocraftonline/heroes/characters/skill/skills/SkillForcePull.java
@@ -47,7 +47,7 @@ public class SkillForcePull extends TargettedSkill {
     public SkillResult use(Hero hero, final LivingEntity target, String[] args) {
         Player player = hero.getPlayer();
 
-        int damage = SkillConfigManager.getUseSetting(hero, this, SkillSetting.DAMAGE, 0, false);
+        double damage = SkillConfigManager.getUseSetting(hero, this, SkillSetting.DAMAGE, 0, false);
         if (damage > 0) {
             addSpellTarget(target, hero);
             damageEntity(target, player, damage, DamageCause.MAGIC);
